// src/structurizr.schema.d.ts
export interface components {
    schemas: {
        /** @description Represents a Structurizr workspace, which is a wrapper for a software architecture model, views, and documentation. */
        Workspace: {
            /**
             * Format: long
             * @description The workspace ID.
             */
            id?: number;
            /** @description The name of the workspace. */
            name?: string;
            /** @description A short description of the workspace. */
            description?: string;
            /** @description A version number for the workspace. */
            version?: string;
            /** @description The thumbnail associated with the workspace; a Base64 encoded PNG file as a data URI (data:image/png;base64). */
            thumbnail?: string;
            /** @description The last modified date, in ISO 8601 format (e.g. "2018-09-08T12:40:03Z"). */
            lastModifiedDate?: string;
            /** @description A string identifying the user who last modified the workspace (e.g. an e-mail address or username). */
            lastModifiedUser?: string;
            /** @description A string identifying the agent that was last used to modify the workspace (e.g. "structurizr-java/1.2.0"). */
            lastModifiedAgent?: string;
            model?: components["schemas"]["Model"];
            views?: components["schemas"]["Views"];
            documentation?: components["schemas"]["Documentation"];
            configuration?: components["schemas"]["WorkspaceConfiguration"];
            /** @description A set of arbitrary name-value properties. */
            properties?: {
                [key: string]: unknown;
            };
        };
        /** @description A software architecture model. */
        Model: {
            /** @description The enterprise associated with this model. */
            enterprise?: {
                /** @description The name of the enterprise. */
                name?: string;
            };
            people?: components["schemas"]["Person"][];
            softwareSystems?: components["schemas"]["SoftwareSystem"][];
            deploymentNodes?: components["schemas"]["DeploymentNode"][];
            /** @description A set of arbitrary name-value properties. */
            properties?: {
                [key: string]: unknown;
            };
        };
        /** @description A person who uses a software system. */
        Person: {
            /** @description The ID of this person in the model. */
            id?: string;
            /** @description The name of this person. */
            name?: string;
            /** @description A short description of this person. */
            description?: string;
            /** @description A comma separated list of tags associated with this person. */
            tags?: string;
            /** @description The URL where more information about this element can be found. */
            url?: string;
            /**
             * @description The location of this person.
             * @enum {string}
             */
            location?: "External" | "Internal" | "Unspecified";
            /** @description The name of the group in which this person should be included in. */
            group?: string;
            /** @description A set of arbitrary name-value properties. */
            properties?: {
                [key: string]: unknown;
            };
            /** @description The set of perspectives associated with this element. */
            perspectives?: components["schemas"]["Perspective"][];
            /** @description The set of relationships from this person to other elements. */
            relationships?: components["schemas"]["Relationship"][];
        };
        /** @description A software system. */
        SoftwareSystem: {
            /** @description The ID of this software system in the model. */
            id?: string;
            /** @description The name of this software system. */
            name?: string;
            /** @description A short description of this software system. */
            description?: string;
            /**
             * @description The location of this software system.
             * @enum {string}
             */
            location?: "External" | "Internal" | "Unspecified";
            /** @description A comma separated list of tags associated with this software system. */
            tags?: string;
            /** @description The URL where more information about this element can be found. */
            url?: string;
            /** @description The set of containers within this software system. */
            containers?: components["schemas"]["Container"][];
            /** @description The name of the group in which this software system should be included in. */
            group?: string;
            /** @description A set of arbitrary name-value properties. */
            properties?: {
                [key: string]: unknown;
            };
            /** @description The set of perspectives associated with this element. */
            perspectives?: components["schemas"]["Perspective"][];
            /** @description The set of relationships from this software system to other elements. */
            relationships?: components["schemas"]["Relationship"][];
            documentation?: components["schemas"]["Documentation"];
        };
        /** @description A container (something that can execute code or host data). */
        Container: {
            /** @description The ID of this container in the model. */
            id?: string;
            /** @description The name of this container. */
            name?: string;
            /** @description A short description of this container. */
            description?: string;
            /** @description The technology associated with this container (e.g. Apache Tomcat). */
            technology?: string;
            /** @description A comma separated list of tags associated with this container. */
            tags?: string;
            /** @description The URL where more information about this element can be found. */
            url?: string;
            /** @description The set of components within this container. */
            components?: components["schemas"]["Component"][];
            /** @description The name of the group in which this container should be included in. */
            group?: string;
            /** @description A set of arbitrary name-value properties. */
            properties?: {
                [key: string]: unknown;
            };
            /** @description The set of perspectives associated with this element. */
            perspectives?: components["schemas"]["Perspective"][];
            /** @description The set of relationships from this container to other elements. */
            relationships?: components["schemas"]["Relationship"][];
            documentation?: components["schemas"]["Documentation"];
        };
        /** @description A component (a grouping of related functionality behind an interface that runs inside a container). */
        Component: {
            /** @description The ID of this component in the model. */
            id?: string;
            /** @description The name of this component. */
            name?: string;
            /** @description A short description of this component. */
            description?: string;
            /** @description The technology associated with this component (e.g. Spring Bean). */
            technology?: string;
            /** @description A comma separated list of tags associated with this component. */
            tags?: string;
            /** @description The URL where more information about this element can be found. */
            url?: string;
            /** @description The name of the group in which this component should be included in. */
            group?: string;
            /** @description A set of arbitrary name-value properties. */
            properties?: {
                [key: string]: unknown;
            };
            /** @description The set of perspectives associated with this element. */
            perspectives?: components["schemas"]["Perspective"][];
            /** @description The set of relationships from this component to other elements. */
            relationships?: components["schemas"]["Relationship"][];
            documentation?: components["schemas"]["Documentation"];
        };
// ...
        Relationship: {
            /** @description The ID of this relationship in the model. */
            id?: string;
            /** @description A short description of this relationship. */
            description?: string;
            /** @description A comma separated list of tags associated with this relationship. */
            tags?: string;
            /** @description The URL where more information about this relationship can be found. */
            url?: string;
            /** @description A set of arbitrary name-value properties. */
            properties?: {
                [key: string]: unknown;
            };
            /** @description The set of perspectives associated with this relationship. */
            perspectives?: components["schemas"]["Perspective"][];
            /** @description The ID of the source element. */
            sourceId?: string;
            /** @description The ID of the destination element. */
            destinationId?: string;
            /** @description The technology associated with this relationship (e.g. HTTPS, JDBC, etc). */
            technology?: string;
            /**
             * @description The interaction style (synchronous or asynchronous).
             * @enum {string}
             */
            interactionStyle?: "Synchronous" | "Asynchronous";
            /** @description The ID of the container-container relationship upon which this container instance-container instance relationship is based. */
            linkedRelationshipId?: string;
        };
// ...
        SystemLandscapeView: {
            /** @description A unique identifier for this view. */
            key?: string;
            /** @description An integer representing the creation order of this view. */
            order?: number;
            /** @description The title of this view (optional). */
            title?: string;
            /** @description The description of this view. */
            description?: string;
            /** @description A set of arbitrary name-value properties. */
            properties?: {
                [key: string]: unknown;
            };
            /**
             * @description The paper size that should be used to render this view.
             * @enum {string}
             */
            paperSize?: "A6_Portrait" | "A6_Landscape" | "A5_Portrait" | "A5_Landscape" | "A4_Portrait" | "A4_Landscape" | "A3_Portrait" | "A3_Landscape" | "A2_Portrait" | "A2_Landscape" | "A1_Portrait" | "A1_Landscape" | "A0_Portrait" | "A0_Landscape" | "Letter_Portrait" | "Letter_Landscape" | "Legal_Portrait" | "Legal_Landscape" | "Slide_4_3" | "Slide_16_9" | "Slide_16_10";
            dimensions?: components["schemas"]["Dimensions"];
            automaticLayout?: components["schemas"]["AutomaticLayout"];
            /** @description Specifies whether the enterprise boundary (to differentiate internal elements from external elements) should be visible on the resulting diagram. */
            enterpriseBoundaryVisible?: boolean;
            /** @description The set of elements in this views. */
            elements?: components["schemas"]["ElementView"][];
            /** @description The set of relationships in this views. */
            relationships?: components["schemas"]["RelationshipView"][];
            /** @description The set of animation steps (optional). */
            animations?: components["schemas"]["AnimationStep"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}